Fastapi/requirements.txt
aiofiles~=0.8.0
av
click~=8.1.3
cloudpickle
Cython
fastapi~=0.78.0
fvcore
h5py
loguru~=0.6.0
matplotlib
memory_profiler
mock
networkx
numpy~=1.22.4
omegaconf
opencv-python~=4.5.5.64
Pillow==9.0.1
pre-commit==2.20.0
pycocotools
pydantic~=1.9.1
pytest~=7.0.0
python-dateutil
python-multipart~=0.0.5
pyyaml>=3.12
requests~=2.27.1
setuptools~=62.3.4
starlette~=0.19.1
scikit-image
scipy
tensorboardX
torch~=1.11.0
torchaudio
torchgeometry
torchmetrics
torchvision~=0.12.0
tqdm~=4.64.0
typing~=3.7.4.3
uvicorn~=0.17.6

